Conflicts:
	apps/include/matrix_engine/engine/simple_engine.h
	engine/src/alg/detector/vehicle_caffe_detector.cpp
	engine/src/engine/simple_engine.h
	engine/src/processor/config_filter.cpp

+++ engine/src/model/rank_feature.h
@@ -75,8 +75,8 @@ protected:
 
 class CarRankFeature final: public RankFeature {
 public:
-    int width_;
-    int height_;
+    ushort width_;
+    ushort height_;
     cv::Mat descriptor_;
     cv::Mat position_;
